abstract tokenizers
